+++ kernel/src/arch/x86_64/mod.rs
@@ -1,0 +1,4 @@
+pub mod descriptors;
+pub mod exceptions;
+pub mod qemu;
+pub mod test;